# Improve performance of manyToManyDescriptor