--- a/src/test/testIntegrations.js
+++ b/src/test/testIntegrations.js
@@ -841,7 +841,7 @@ describe('Many-to-many relationship performance', () => {
         const end = new Date().getTime();
         const tookSeconds = (end - start) / 1000;
         console.log(`Adding ${childAmount} relations took ${tookSeconds}s`);
-        expect(tookSeconds).toBeLessThanOrEqual(3);
+        expect(tookSeconds).toBeLessThanOrEqual(process.env.TRAVIS ? 10 : 3);
     });
 
     it('queries many-to-many relationships in acceptable time', () => {
@@ -867,7 +867,7 @@ describe('Many-to-many relationship performance', () => {
         const end = new Date().getTime();
         const tookSeconds = (end - start) / 1000;
         console.log(`Performing ${queryCount} queries took ${tookSeconds}s`);
-        expect(tookSeconds).toBeLessThanOrEqual(3);
+        expect(tookSeconds).toBeLessThanOrEqual(process.env.TRAVIS ? 10 : 3);
     });
 
     it('removes many-to-many relationships in acceptable time', () => {
@@ -893,6 +893,6 @@ describe('Many-to-many relationship performance', () => {
         const end = new Date().getTime();
         const tookSeconds = (end - start) / 1000;
         console.log(`Removing ${removeCount} relations took ${tookSeconds}s`);
-        expect(tookSeconds).toBeLessThanOrEqual(3);
+        expect(tookSeconds).toBeLessThanOrEqual(process.env.TRAVIS ? 10 : 3);
     });
 });