Use Set structures to speed up collection difference and intersection operations, particularly in the case of querying for many-to-many relationships when there are large numbers of the "toModel".

**Before**

![before](https://user-images.githubusercontent.com/35026/31795196-e741e56a-b4f2-11e7-8777-7a2c878beacd.png)

**After**

![after](https://user-images.githubusercontent.com/35026/31795199-e8f419f0-b4f2-11e7-91cc-155c96fb666c.png)

**Summary**

|        | Before  | After  |
|:------:|---------|--------|
| Create |   3.46s | 2.049s |
| Read   | 36.121s | 1.617s |
| Delete |  3.161s | 1.984s |