@tiii 2017-10-20
Seems like Travis is not fast enought for the 3 seconds you suggested 😉 

Isn't Set an es6 feature? So redux-orm users must likely have to install a polyfill in their apps? I'm fine with that, since we gain speed for most users, but it has to be documented somewhere.

Nonetheless: Great work! 👍