@ahutchings 2017-10-20
Yeah, it's probably best to pull all of the benchmark/performance tests out of the main "npm test" script since Travis isn't really designed for perf testing. I'll make that change.

Set is ES6, and looks pretty widely supported (node 4+, latest browsers), but this should be released in a major version to ensure it doesn't break any users. I'll add docs for the polyfill.