@tiii 2017-10-20
Extracting benchmark tests seems like a good idea, but I would still run them and just adjust set the limit higher (or use a modifier for running them on travis?). Those limits are always dependent on the computer running them. It's still better than not testing for the performance.