@ahutchings 2017-10-20
Updated to add docs for the polyfill, and increased the benchmark expectations to 10 seconds when running in Travis.