@tiii 2018-01-01
@ahutchings for some reason the tests now take even longer as before and therefore fail on master. Could you take a look?