@ahutchings 2018-01-01
I noticed that a PR I submitted a few minutes ago is passing (#181). Maybe Travis was swamped when the last two PRs were merged into master? We could bump up the timing thresholds a bit to give the tests some more clearance, but in this case Travis was so much slower than normal that we'd have to bump the thresholds so far that they wouldn't be useful anymore.

I wonder if there is some dedicated performance testing service that we could use for perf tests instead of Travis. I'll do some research.