@tiii 2018-01-04
Thanks for checking it out! Since the current master is green again, I would suggest we just keep it that way and hope travis won't choke on a regular basis.

I think we can speed up the performance tests a bit more by running them without coverage (though this may only gain a few ms)? IMHO those shouldn't contribute to the test coverage anyways.